requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
twilio==8.11.0
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
google-generativeai==0.3.2
python-dotenv==1.0.0
websockets==12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
